refactor: split onboard entry render
